divide now returns double

+++ src/main/java/com/example/springdocker/model/MyMathCalc.java
@@ -8,8 +8,8 @@ public class MyMathCalc {
     public int multiply(int a, int b){
         return a*b;
     }
-
-    public float divide(float a, float b){
+  
+    public double divide(double a, double b){
         return a/b;
     }
 }
